// frontend/src/lib/api/api.interfaces.d.ts
export interface IGetUserResponse {
    email: string;
    full_name: string;
    disabled: boolean;
    auth_method: 'email' | 'google';
    photo: string;
    id: string;
    has_api_key: boolean;
}
// ...
export interface Member {
    id: string;
    organization_id: string;
    role: OrganizationRoleEnum;
    user: Omit<IGetUserResponse, 'auth_method'>;
    created_at: Date;
    updated_at: Date | null;
}
// ...
export interface OrbitMember extends Omit<Member, 'organization_id' | 'role'> {
    orbit_id: string;
    role: OrbitRoleEnum;
}